Refactored to split out platform-agnostic stuff into ChatPlatformController from SlackController - this should give us a nice place to fill in the MS Teams functionality
