+++ app/controllers/SlackController.scala
@@ -1,18 +1,17 @@
 package controllers
 
 import com.google.inject.Provider
-import com.mohiva.play.silhouette.api.{LoginInfo, Silhouette}
+import com.mohiva.play.silhouette.api.Silhouette
 import javax.inject.Inject
 import json.Formatting._
 import models.accounts.slack.botprofile.SlackBotProfile
 import models.behaviors.behaviorgroupversion.BehaviorGroupVersion
 import models.behaviors.behaviorversion.{Normal, Threaded}
-import models.behaviors.builtins.DisplayHelpBehavior
 import models.behaviors.conversations.conversation.Conversation
 import models.behaviors.events.MessageActionConstants._
 import models.behaviors.events._
 import models.behaviors.events.slack._
-import models.behaviors.{ActionChoice, SimpleTextResult}
+import models.behaviors.{ActionChoice, BotResult, SimpleTextResult}
 import models.help.HelpGroupSearchValue
 import models.silhouette.EllipsisEnv
 import play.api.data.Form
@@ -36,12 +35,13 @@ class SlackController @Inject() (
                                   val assetsProvider: Provider[RemoteAssets],
                                   val environment: Environment,
                                   implicit val ec: ExecutionContext
-                                ) extends EllipsisController {
+                                ) extends EllipsisController with ChatPlatformController {
 
-  val dataService = services.dataService
+  override type ActionsTriggeredInfoType = ActionsTriggeredInfo
+  override type BotProfileType = SlackBotProfile
+
   val configuration = services.configuration
   val lambdaService = services.lambdaService
-  val cacheService = services.cacheService
   val ws = services.ws
   val botResultService = services.botResultService
   implicit val actorSystem = services.actorSystem
@@ -90,6 +90,25 @@ class SlackController @Inject() (
         Forbidden("Bad token")
       }
     })
+  }
+
+  def sendEphemeralMessageFor(message: String, slackTeamId: String, slackChannelId: String, maybeThreadTs: Option[String], slackUserId: String): Future[Unit] = {
+    for {
+      maybeProfile <- dataService.slackBotProfiles.allForSlackTeamId(slackTeamId).map(_.headOption)
+      _ <- (for {
+        profile <- maybeProfile
+      } yield {
+        services.slackApiService.clientFor(profile).postEphemeralMessage(message, slackChannelId, maybeThreadTs, slackUserId)
+      }).getOrElse {
+        Future.successful({})
+      }
+    } yield {}
+  }
+
+  def sendEphemeralMessage(message: String, info: MessageRequestInfo): Future[Unit] = {
+    info.slackTeamIdsForBots.headOption.map { slackTeamId =>
+      sendEphemeralMessageFor(message, slackTeamId, info.channel, info.maybeThreadTs, info.userId)
+    }.getOrElse(Future.successful({}))
   }
 
   trait EventInfo {
@@ -588,7 +607,261 @@ class SlackController @Inject() (
                                    token: String,
                                    original_message: Option[OriginalMessageInfo],
                                    response_url: String
-                                 ) extends RequestInfo {
+                                 ) extends RequestInfo with ActionsTriggeredInfoInterface {
+
+    val channelId: String = channel.id
+    val teamIdForContext: String = slackTeamIdForBot
+    val teamIdForUserForContext: String = slackTeamIdForUser
+    val userIdForContext: String = user.id
+
+    val contextName: String = Conversation.SLACK_CONTEXT
+
+    def onEvent(event: Event): Future[Unit] = slackEventService.onEvent(event)
+
+    def sendResultWithNewEvent(
+                                description: String,
+                                getEventualMaybeResult: MessageEvent => Future[Option[BotResult]],
+                                botProfile: BotProfileType,
+                                beQuiet: Boolean
+                              ): Future[Option[String]] = {
+      dataService.slackBotProfiles.sendResultWithNewEvent(
+        "help index",
+        getEventualMaybeResult,
+        botProfile,
+        channel.id,
+        user.id,
+        message_ts,
+        maybeOriginalMessageThreadId,
+        isEphemeral,
+        Some(response_url),
+        beQuiet = false
+      )
+
+    }
+
+    def inputChoiceResultFor(value: String)(implicit request: Request[AnyContent]): Future[Unit] = {
+      for {
+        maybeProfile <- dataService.slackBotProfiles.allForSlackTeamId(team.id).map(_.headOption)
+        maybeSlackMessage <- maybeProfile.map { profile =>
+          SlackMessage.fromFormattedText(value, profile, slackEventService).map(Some(_))
+        }.getOrElse(Future.successful(None))
+        _ <- (for {
+          profile <- maybeProfile
+          slackMessage <- maybeSlackMessage
+        } yield {
+          slackEventService.onEvent(SlackMessageEvent(
+            SlackEventContext(
+              profile,
+              channel.id,
+              maybeOriginalMessageThreadId,
+              user.id
+            ),
+            slackMessage,
+            None,
+            message_ts,
+            None,
+            isUninterruptedConversation = false,
+            isEphemeral,
+            Some(response_url),
+            beQuiet = false
+          ))
+        }).getOrElse {
+          Future.successful({})
+        }
+      } yield {}
+    }
+
+    def sendEphemeralMessage(message: String): Future[Unit] = {
+      sendEphemeralMessageFor(message, slackTeamIdForBot, channel.id, maybeOriginalMessageThreadId, user.id)
+    }
+
+    def updateActionsMessageFor(
+                                 maybeResultText: Option[String],
+                                 shouldRemoveActions: Boolean
+                               ): Future[Unit] = {
+      original_message.map { originalMessage =>
+        val maybeOriginalColor = originalMessage.attachments.headOption.flatMap(_.color)
+        val newAttachment = AttachmentInfo(maybeResultText, None, None, Some(Seq("text")), Some(callback_id), color = maybeOriginalColor, footer = maybeResultText)
+        val originalAttachmentsToUse = if (shouldRemoveActions) {
+          originalMessage.attachments.map(ea => ea.copy(actions = None))
+        } else {
+          originalMessage.attachments
+        }
+        val updated = originalMessage.copy(attachments = originalAttachmentsToUse :+ newAttachment)
+        for {
+          maybeProfile <- dataService.slackBotProfiles.allForSlackTeamId(team.id).map(_.headOption)
+          _ <- (for {
+            profile <- maybeProfile
+          } yield {
+            ws.
+              url("https://slack.com/api/chat.update").
+              withHttpHeaders(HeaderNames.ACCEPT -> MimeTypes.JSON).
+              post(Map(
+                "token" -> Seq(profile.token),
+                "channel" -> Seq(channel.id),
+                "text" -> Seq(updated.text),
+                "attachments" -> Seq(Json.prettyPrint(Json.toJson(updated.attachments))),
+                "as_user" -> Seq("true"),
+                "ts" -> Seq(message_ts),
+                "user" -> Seq(user.id)
+              ))
+          }).getOrElse {
+            Future.successful({})
+          }
+        } yield {}
+      }.getOrElse {
+        Future.successful({})
+      }
+    }
+
+    private def cannotBeTriggeredMessageFor(
+                                             actionChoice: ActionChoice,
+                                             maybeGroupVersion: Option[BehaviorGroupVersion],
+                                             maybeChoiceSlackUserId: Option[String]
+                                           ): String = {
+      if (actionChoice.areOthersAllowed) {
+        val teamText = maybeGroupVersion.map { bgv => s" ${bgv.team.name}"}.getOrElse("")
+        s"Only members of the${teamText} team can make this choice"
+      } else {
+        maybeChoiceSlackUserId.map { choiceSlackUserId =>
+          s"Only <@${choiceSlackUserId}> can make this choice"
+        }.getOrElse("You are not allowed to make this choice")
+      }
+    }
+
+    def sendCannotBeTriggeredFor(
+                                  actionChoice: ActionChoice,
+                                  maybeGroupVersion: Option[BehaviorGroupVersion]
+                                ): Future[Unit] = {
+      for {
+        maybeChoiceSlackUserId <- maybeSlackUserIdForActionChoice(actionChoice)
+        _ <- {
+          val msg = cannotBeTriggeredMessageFor(actionChoice, maybeGroupVersion, maybeChoiceSlackUserId)
+          sendEphemeralMessage(msg)
+        }
+      } yield {}
+    }
+
+    def processTriggerableAndActiveActionChoice(
+                                                 actionChoice: ActionChoice,
+                                                 maybeGroupVersion: Option[BehaviorGroupVersion],
+                                                 botProfile: SlackBotProfile,
+                                                 maybeInstantResponseTs: Option[String]
+                                               ): Future[Unit] = {
+      for {
+        maybeThreadIdToUse <- maybeOriginalMessageThreadId.map(tid => Future.successful(Some(tid))).getOrElse {
+          dataService.behaviorVersions.findWithoutAccessCheck(actionChoice.originatingBehaviorVersionId).map { maybeOriginatingBehaviorVersion =>
+            if (maybeOriginatingBehaviorVersion.exists(_.responseType == Threaded)) {
+              maybeInstantResponseTs.orElse(original_message.map(_.ts))
+            } else {
+              None
+            }
+          }
+        }
+        _ <- dataService.slackBotProfiles.sendResultWithNewEvent(
+          s"run action named ${actionChoice.actionName}",
+          event => for {
+            maybeBehaviorVersion <- maybeGroupVersion.map { groupVersion =>
+              dataService.behaviorVersions.findByName(actionChoice.actionName, groupVersion)
+            }.getOrElse(Future.successful(None))
+            params <- maybeBehaviorVersion.map { behaviorVersion =>
+              dataService.behaviorParameters.allFor(behaviorVersion)
+            }.getOrElse(Future.successful(Seq()))
+            invocationParams <- Future.successful(actionChoice.argumentsMap.flatMap { case(name, value) =>
+              params.find(_.name == name).map { param =>
+                (AWSLambdaConstants.invocationParamFor(param.rank - 1), value)
+              }
+            })
+            maybeResponse <- maybeBehaviorVersion.map { behaviorVersion =>
+              dataService.behaviorResponses.buildFor(
+                event,
+                behaviorVersion,
+                invocationParams,
+                None,
+                None,
+                None,
+                userExpectsResponse = true
+              ).map(Some(_))
+            }.getOrElse(Future.successful(None))
+            maybeResult <- maybeResponse.map { response =>
+              response.result.map(Some(_))
+            }.getOrElse(Future.successful(None))
+          } yield maybeResult,
+          botProfile,
+          channel.id,
+          user.id,
+          message_ts,
+          maybeThreadIdToUse,
+          isEphemeral,
+          Some(response_url),
+          actionChoice.shouldBeQuiet
+        )
+      } yield {}
+    }
+
+    def instantBackgroundResponse(responseText: String, permission: ActionPermission): Future[Option[String]] = {
+      val trimmed = responseText.trim.replaceAll("(^\\u00A0|\\u00A0$)", "")
+      if (trimmed.isEmpty) {
+        Future.successful(None)
+      } else {
+        for {
+          maybeBotProfile <- maybeBotProfile
+          maybeTs <- maybeBotProfile.map { botProfile =>
+            dataService.slackBotProfiles.sendResultWithNewEvent(
+              "Message acknowledging response to Slack action",
+              slackMessageEvent => for {
+                maybeConversation <- slackMessageEvent.maybeOngoingConversation(dataService)
+              } yield {
+                Some(SimpleTextResult(
+                  slackMessageEvent,
+                  maybeConversation,
+                  s"_${trimmed}_",
+                  responseType = Normal,
+                  shouldInterrupt = false
+                ))
+              },
+              botProfile,
+              channel.id,
+              user.id,
+              message_ts,
+              maybeOriginalMessageThreadId,
+              isEphemeral || permission.beQuiet,
+              Some(response_url),
+              beQuiet = false
+            )
+          }.getOrElse(Future.successful(None))
+        } yield maybeTs
+      }
+    }
+
+    def result(maybeResultText: Option[String], permission: ActionPermission): Result = {
+
+      // respond immediately by sending a new message
+      val instantResponse = maybeResultText.map(t => instantBackgroundResponse(t, permission)).getOrElse(Future.successful(None))
+      permission.runInBackground(instantResponse)
+
+      val updated = if (permission.shouldRemoveActions) {
+        original_message.map { originalMessage =>
+          val maybeOriginalColor = originalMessage.attachments.headOption.flatMap(_.color)
+          val newAttachment = AttachmentInfo(
+            maybeResultText,
+            title = None,
+            text = None,
+            Some(Seq("text")),
+            Some(callback_id),
+            color = maybeOriginalColor,
+            footer = Some("✔︎︎ OK")
+          )
+          val attachments = originalMessage.attachments.map(ea => ea.copy(actions = None))
+          originalMessage.copy(attachments = attachments :+ newAttachment)
+        }
+      } else {
+        original_message
+      }
+      updated.map { u =>
+        Ok(Json.toJson(u))
+      }.getOrElse(Ok(""))
+    }
 
     val maybeOriginalMessageThreadId: Option[String] = {
       val maybeThreadId = original_message.flatMap(_.thread_ts)
@@ -749,6 +1022,21 @@ class SlackController @Inject() (
       }.getOrElse(Future.successful(false))
     }
 
+    def isIncorrectTeam(botProfile: BotProfileType): Future[Boolean] = {
+      for {
+        maybeUser <- dataService.users.ensureUserFor(loginInfo, botProfile.teamId).map(Some(_))
+        isAdmin <- maybeUser.map { user =>
+          dataService.users.isAdmin(user)
+        }.getOrElse(Future.successful(false))
+        maybeAttemptingSlackTeamIds <- maybeUser.map { user =>
+          dataService.users.maybeSlackTeamIdsFor(user)
+        }.getOrElse(Future.successful(None))
+      } yield {
+        val isSameTeam = maybeAttemptingSlackTeamIds.exists(_.contains(botProfile.teamIdForContext))
+        !team.isEnterpriseGrid && !isSameTeam && !isAdmin
+      }
+    }
+
     private def originalMessageActions: Seq[ActionInfo] = {
       this.original_message.map { msg =>
         msg.attachments.flatMap(_.actions).flatten
@@ -774,11 +1062,16 @@ class SlackController @Inject() (
       }
       maybeAction.map(_.text)
     }
+
+    def formattedUserFor(permission: ActionPermission): String = {
+      if (channel.isDirectMessage || permission.beQuiet) {
+        "You"
+      } else {
+        s"<@${user.id}>"
+      }
+    }
+
   }
-
-  case class ConfirmContinueConversationResponse(shouldContinue: Boolean, conversationId: String, userId: String)
-
-  case class StopConversationResponse(conversationId: String, userId: String)
 
   private val actionForm = Form(
     "payload" -> nonEmptyText
@@ -807,877 +1100,12 @@ class SlackController @Inject() (
 
   implicit val actionsTriggeredReads = Json.reads[ActionsTriggeredInfo]
 
-  private def sendEphemeralMessage(message: String, slackTeamId: String, slackChannelId: String, maybeThreadTs: Option[String], slackUserId: String): Future[Unit] = {
-    for {
-      maybeProfile <- dataService.slackBotProfiles.allForSlackTeamId(slackTeamId).map(_.headOption)
-      _ <- (for {
-        profile <- maybeProfile
-      } yield {
-        services.slackApiService.clientFor(profile).postEphemeralMessage(message, slackChannelId, maybeThreadTs, slackUserId)
-      }).getOrElse {
-        Future.successful({})
-      }
-    } yield {}
-  }
-
-  private def sendEphemeralMessage(message: String, info: ActionsTriggeredInfo): Future[Unit] = {
-    sendEphemeralMessage(message, info.slackTeamIdForBot, info.channel.id, info.maybeOriginalMessageThreadId, info.user.id)
-  }
-
-  private def sendEphemeralMessage(message: String, info: MessageRequestInfo): Future[Unit] = {
-    info.slackTeamIdsForBots.headOption.map { slackTeamId =>
-      sendEphemeralMessage(message, slackTeamId, info.channel, info.maybeThreadTs, info.userId)
-    }.getOrElse(Future.successful({}))
-  }
-
-  private def updateActionsMessageFor(
-                                       info: ActionsTriggeredInfo,
-                                       maybeResultText: Option[String],
-                                       shouldRemoveActions: Boolean
-                                     ): Future[Unit] = {
-    info.original_message.map { originalMessage =>
-      val maybeOriginalColor = originalMessage.attachments.headOption.flatMap(_.color)
-      val newAttachment = AttachmentInfo(maybeResultText, None, None, Some(Seq("text")), Some(info.callback_id), color = maybeOriginalColor, footer = maybeResultText)
-      val originalAttachmentsToUse = if (shouldRemoveActions) {
-        originalMessage.attachments.map(ea => ea.copy(actions = None))
-      } else {
-        originalMessage.attachments
-      }
-      val updated = originalMessage.copy(attachments = originalAttachmentsToUse :+ newAttachment)
-      for {
-        maybeProfile <- dataService.slackBotProfiles.allForSlackTeamId(info.team.id).map(_.headOption)
-        _ <- (for {
-          profile <- maybeProfile
-        } yield {
-          ws.
-            url("https://slack.com/api/chat.update").
-            withHttpHeaders(HeaderNames.ACCEPT -> MimeTypes.JSON).
-            post(Map(
-              "token" -> Seq(profile.token),
-              "channel" -> Seq(info.channel.id),
-              "text" -> Seq(updated.text),
-              "attachments" -> Seq(Json.prettyPrint(Json.toJson(updated.attachments))),
-              "as_user" -> Seq("true"),
-              "ts" -> Seq(info.message_ts),
-              "user" -> Seq(info.user.id)
-            ))
-        }).getOrElse {
-          Future.successful({})
-        }
-      } yield {}
-    }.getOrElse {
-      Future.successful({})
-    }
-  }
-
-  private def inputChoiceResultFor(value: String, info: ActionsTriggeredInfo)(implicit request: Request[AnyContent]): Future[Unit] = {
-    for {
-      maybeProfile <- dataService.slackBotProfiles.allForSlackTeamId(info.team.id).map(_.headOption)
-      maybeSlackMessage <- maybeProfile.map { profile =>
-        SlackMessage.fromFormattedText(value, profile, slackEventService).map(Some(_))
-      }.getOrElse(Future.successful(None))
-      _ <- (for {
-        profile <- maybeProfile
-        slackMessage <- maybeSlackMessage
-      } yield {
-        slackEventService.onEvent(SlackMessageEvent(
-          SlackEventContext(
-            profile,
-            info.channel.id,
-            info.maybeOriginalMessageThreadId,
-            info.user.id
-          ),
-          slackMessage,
-          None,
-          info.message_ts,
-          None,
-          isUninterruptedConversation = false,
-          info.isEphemeral,
-          Some(info.response_url),
-          beQuiet = false
-        ))
-      }).getOrElse {
-        Future.successful({})
-      }
-    } yield {}
-  }
-
   private def maybeSlackUserIdForActionChoice(actionChoice: ActionChoice): Future[Option[String]] = {
     dataService.users.find(actionChoice.userId).flatMap { maybeUser =>
       maybeUser.map { user =>
         dataService.linkedAccounts.maybeSlackUserIdFor(user)
       }.getOrElse(Future.successful(None))
     }
-  }
-
-  private def cannotBeTriggeredMessageFor(
-                                           actionChoice: ActionChoice,
-                                           maybeGroupVersion: Option[BehaviorGroupVersion],
-                                           maybeChoiceSlackUserId: Option[String]
-                                         ): String = {
-    if (actionChoice.areOthersAllowed) {
-      val teamText = maybeGroupVersion.map { bgv => s" ${bgv.team.name}"}.getOrElse("")
-      s"Only members of the${teamText} team can make this choice"
-    } else {
-      maybeChoiceSlackUserId.map { choiceSlackUserId =>
-        s"Only <@${choiceSlackUserId}> can make this choice"
-      }.getOrElse("You are not allowed to make this choice")
-    }
-  }
-
-  private def sendCannotBeTriggeredFor(
-                                        actionChoice: ActionChoice,
-                                        maybeGroupVersion: Option[BehaviorGroupVersion],
-                                        info: ActionsTriggeredInfo
-                                      ): Future[Unit] = {
-    for {
-      maybeChoiceSlackUserId <- maybeSlackUserIdForActionChoice(actionChoice)
-      _ <- {
-        val msg = cannotBeTriggeredMessageFor(actionChoice, maybeGroupVersion, maybeChoiceSlackUserId)
-        sendEphemeralMessage(msg, info)
-      }
-    } yield {}
-  }
-
-  private def processTriggerableAndActiveActionChoice(
-                                                       actionChoice: ActionChoice,
-                                                       maybeGroupVersion: Option[BehaviorGroupVersion],
-                                                       info: ActionsTriggeredInfo,
-                                                       botProfile: SlackBotProfile,
-                                                       maybeInstantResponseTs: Option[String]
-                                                     ): Future[Unit] = {
-    for {
-      maybeThreadIdToUse <- info.maybeOriginalMessageThreadId.map(tid => Future.successful(Some(tid))).getOrElse {
-        dataService.behaviorVersions.findWithoutAccessCheck(actionChoice.originatingBehaviorVersionId).map { maybeOriginatingBehaviorVersion =>
-          if (maybeOriginatingBehaviorVersion.exists(_.responseType == Threaded)) {
-            maybeInstantResponseTs.orElse(info.original_message.map(_.ts))
-          } else {
-            None
-          }
-        }
-      }
-      _ <- dataService.slackBotProfiles.sendResultWithNewEvent(
-        s"run action named ${actionChoice.actionName}",
-        event => for {
-          maybeBehaviorVersion <- maybeGroupVersion.map { groupVersion =>
-            dataService.behaviorVersions.findByName(actionChoice.actionName, groupVersion)
-          }.getOrElse(Future.successful(None))
-          params <- maybeBehaviorVersion.map { behaviorVersion =>
-            dataService.behaviorParameters.allFor(behaviorVersion)
-          }.getOrElse(Future.successful(Seq()))
-          invocationParams <- Future.successful(actionChoice.argumentsMap.flatMap { case(name, value) =>
-            params.find(_.name == name).map { param =>
-              (AWSLambdaConstants.invocationParamFor(param.rank - 1), value)
-            }
-          })
-          maybeResponse <- maybeBehaviorVersion.map { behaviorVersion =>
-            dataService.behaviorResponses.buildFor(
-              event,
-              behaviorVersion,
-              invocationParams,
-              None,
-              None,
-              None,
-              userExpectsResponse = true
-            ).map(Some(_))
-          }.getOrElse(Future.successful(None))
-          maybeResult <- maybeResponse.map { response =>
-            response.result.map(Some(_))
-          }.getOrElse(Future.successful(None))
-        } yield maybeResult,
-        botProfile,
-        info.channel.id,
-        info.user.id,
-        info.message_ts,
-        maybeThreadIdToUse,
-        info.isEphemeral,
-        Some(info.response_url),
-        actionChoice.shouldBeQuiet
-      )
-    } yield {}
-  }
-
-  trait ActionPermission {
-
-    val info: ActionsTriggeredInfo
-    val shouldRemoveActions: Boolean
-    val maybeResultText: Option[String]
-    val beQuiet: Boolean = false
-    def slackUser: String = if (info.channel.isDirectMessage || beQuiet) {
-      "You"
-    } else {
-      s"<@${info.user.id}>"
-    }
-    implicit val request: Request[AnyContent]
-
-    def instantBackgroundResponse(responseText: String): Future[Option[String]] = {
-      val trimmed = responseText.trim.replaceAll("(^\\u00A0|\\u00A0$)", "")
-      if (trimmed.isEmpty) {
-        Future.successful(None)
-      } else {
-        for {
-          maybeBotProfile <- info.maybeBotProfile
-          maybeTs <- maybeBotProfile.map { botProfile =>
-            dataService.slackBotProfiles.sendResultWithNewEvent(
-              "Message acknowledging response to Slack action",
-              slackMessageEvent => for {
-                maybeConversation <- slackMessageEvent.maybeOngoingConversation(dataService)
-              } yield {
-                Some(SimpleTextResult(
-                  slackMessageEvent,
-                  maybeConversation,
-                  s"_${trimmed}_",
-                  responseType = Normal,
-                  shouldInterrupt = false
-                ))
-              },
-              botProfile,
-              info.channel.id,
-              info.user.id,
-              info.message_ts,
-              info.maybeOriginalMessageThreadId,
-              info.isEphemeral || beQuiet,
-              Some(info.response_url),
-              beQuiet = false
-            )
-          }.getOrElse(Future.successful(None))
-        } yield maybeTs
-      }
-    }
-
-    def runInBackground(maybeInstantResponseTs: Future[Option[String]]): Unit
-
-    def result: Result = {
-
-      // respond immediately by sending a new message
-      val instantResponse = maybeResultText.map(instantBackgroundResponse).getOrElse(Future.successful(None))
-      runInBackground(instantResponse)
-
-      val updated = if (shouldRemoveActions) {
-        info.original_message.map { originalMessage =>
-          val maybeOriginalColor = originalMessage.attachments.headOption.flatMap(_.color)
-          val newAttachment = AttachmentInfo(
-            maybeResultText,
-            title = None,
-            text = None,
-            Some(Seq("text")),
-            Some(info.callback_id),
-            color = maybeOriginalColor,
-            footer = Some("✔︎︎ OK")
-          )
-          val attachments = originalMessage.attachments.map(ea => ea.copy(actions = None))
-          originalMessage.copy(attachments = attachments :+ newAttachment)
-        }
-      } else {
-        info.original_message
-      }
-      updated.map { u =>
-        Ok(Json.toJson(u))
-      }.getOrElse(Ok(""))
-    }
-
-  }
-
-  trait ActionPermissionType[T <: ActionPermission] {
-
-    def maybeFor(info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Option[Future[T]]
-
-    def maybeResultFor(info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Option[Future[Result]] = {
-      maybeFor(info, botProfile).map(_.map(_.result))
-    }
-
-  }
-
-  case class ActionChoicePermission(
-                                     actionChoice: ActionChoice,
-                                     info: ActionsTriggeredInfo,
-                                     maybeGroupVersion: Option[BehaviorGroupVersion],
-                                     isActive: Boolean,
-                                     canBeTriggered: Boolean,
-                                     botProfile: SlackBotProfile,
-                                     implicit val request: Request[AnyContent]
-                                   ) extends ActionPermission {
-
-    override val beQuiet: Boolean = actionChoice.shouldBeQuiet
-
-    override val maybeResultText: Option[String] = if (isActive) {
-      Some(s"$slackUser clicked ${actionChoice.label}")
-    } else {
-      Some("This skill has been updated, making these associated actions no longer valid")
-    }
-
-    override val shouldRemoveActions: Boolean = !actionChoice.allowMultipleSelections.exists(identity) && canBeTriggered
-
-    def runInBackground(maybeInstantResponseTs: Future[Option[String]]) = {
-      if (canBeTriggered) {
-        if (isActive) {
-          for {
-            maybeTs <- maybeInstantResponseTs
-            result <- processTriggerableAndActiveActionChoice(actionChoice, maybeGroupVersion, info, botProfile, maybeTs)
-          } yield result
-        }
-      } else {
-        sendCannotBeTriggeredFor(actionChoice, maybeGroupVersion, info)
-      }
-    }
-  }
-
-  object ActionChoicePermission extends ActionPermissionType[ActionChoicePermission] {
-
-    def maybeFor(info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Option[Future[ActionChoicePermission]] = {
-      info.maybeSelectedActionChoice.map { actionChoice =>
-        buildFor(actionChoice, info, botProfile)
-      }
-    }
-
-    def buildFor(actionChoice: ActionChoice, info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Future[ActionChoicePermission] = {
-      for {
-        maybeOriginatingBehaviorVersion <- dataService.behaviorVersions.findWithoutAccessCheck(actionChoice.originatingBehaviorVersionId)
-        maybeGroupVersion <- Future.successful(maybeOriginatingBehaviorVersion.map(_.groupVersion))
-        maybeActiveGroupVersion <- maybeGroupVersion.map { groupVersion =>
-          dataService.behaviorGroupDeployments.maybeActiveBehaviorGroupVersionFor(groupVersion.group, Conversation.SLACK_CONTEXT, info.channel.id)
-        }.getOrElse(Future.successful(None))
-        isActive <- (for {
-          groupVersion <- maybeGroupVersion
-          activeVersion <- maybeActiveGroupVersion
-        } yield {
-          if (groupVersion == activeVersion) {
-            Future.successful(true)
-          } else {
-            dataService.behaviorGroupVersions.haveActionsWithNameAndSameInterface(actionChoice.actionName, groupVersion, activeVersion)
-          }
-        }).getOrElse(Future.successful(false))
-        canBeTriggered <- for {
-          maybeUser <- dataService.users.ensureUserFor(LoginInfo(Conversation.SLACK_CONTEXT, info.user.id), botProfile.teamId).map(Some(_))
-          canBeTriggered <- maybeUser.map { user =>
-            actionChoice.canBeTriggeredBy(user, info.slackTeamIdForUser, botProfile.slackTeamId, dataService)
-          }.getOrElse(Future.successful(false))
-        } yield canBeTriggered
-      } yield ActionChoicePermission(
-        actionChoice,
-        info,
-        maybeActiveGroupVersion,
-        isActive,
-        canBeTriggered,
-        botProfile,
-        request
-      )
-    }
-  }
-
-  trait InputChoicePermission extends ActionPermission {
-    val choice: String
-    val isConversationDone: Boolean
-    val isIncorrectUser: Boolean
-    val maybeResultText = Some(s"$slackUser chose $choice")
-    val shouldRemoveActions = true
-
-    def runInBackground(maybeInstantResponseTs: Future[Option[String]]) = {
-      if (isConversationDone) {
-        updateActionsMessageFor(info, Some(s"This conversation is no longer active"), shouldRemoveActions = true)
-      } else if (isIncorrectUser) {
-        info.maybeUserIdForDataTypeChoice.foreach { correctUserId =>
-          val correctUser = s"<@${correctUserId}>"
-          sendEphemeralMessage(s"Only $correctUser can answer this", info)
-        }
-      } else {
-        inputChoiceResultFor(choice, info)
-      }
-    }
-  }
-
-  case class DataTypeChoicePermission(
-                                    choice: String,
-                                    info: ActionsTriggeredInfo,
-                                    isConversationDone: Boolean,
-                                    implicit val request: Request[AnyContent]
-                                  ) extends InputChoicePermission {
-    val isIncorrectUser: Boolean = info.isIncorrectUserTryingDataTypeChoice
-  }
-
-  object DataTypeChoicePermission extends ActionPermissionType[DataTypeChoicePermission] {
-
-    def maybeFor(info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Option[Future[DataTypeChoicePermission]] = {
-      info.maybeDataTypeChoice.map { choice =>
-        buildFor(choice, info)
-      }
-    }
-
-    def buildFor(choice: String, info: ActionsTriggeredInfo)(implicit request: Request[AnyContent]): Future[DataTypeChoicePermission] = {
-      for {
-        isConversationDone <- info.isForDataTypeChoiceForDoneConversation
-      } yield DataTypeChoicePermission(
-        choice,
-        info,
-        isConversationDone,
-        request
-      )
-    }
-
-  }
-
-  case class YesNoChoicePermission(
-                                    choice: String,
-                                    info: ActionsTriggeredInfo,
-                                    isConversationDone: Boolean,
-                                    implicit val request: Request[AnyContent]
-                                  ) extends InputChoicePermission {
-    val isIncorrectUser: Boolean = info.isIncorrectUserTryingYesNo
-  }
-
-  object YesNoChoicePermission extends ActionPermissionType[YesNoChoicePermission] {
-
-    def maybeFor(info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Option[Future[YesNoChoicePermission]] = {
-      info.maybeYesNoAnswer.map { value =>
-        buildFor(value, info)
-      }
-    }
-
-    def buildFor(value: String, info: ActionsTriggeredInfo)(implicit request: Request[AnyContent]): Future[YesNoChoicePermission] = {
-      for {
-        isConversationDone <- info.isForYesNoForDoneConversation
-      } yield YesNoChoicePermission(
-        value,
-        info,
-        isConversationDone,
-        request
-      )
-    }
-
-  }
-
-  trait HelpPermission extends ActionPermission {
-    val shouldRemoveActions: Boolean = false
-    val isIncorrectTeam: Boolean
-    val botProfile: SlackBotProfile
-
-    def runForCorrectTeam: Unit
-
-    def runInBackground(maybeInstantResponseTs: Future[Option[String]]): Unit = {
-      if (isIncorrectTeam) {
-        dataService.teams.find(botProfile.teamId).flatMap { maybeTeam =>
-          val teamText = maybeTeam.map { team => s" ${team.name}"}.getOrElse("")
-          val msg = s"Only members of the${teamText} team can make this choice"
-          sendEphemeralMessage(msg, info)
-        }
-      } else {
-        runForCorrectTeam
-      }
-    }
-
-  }
-
-  trait HelpPermissionType[T <: HelpPermission, V] extends ActionPermissionType[T] {
-
-    def maybeValueFor(info: ActionsTriggeredInfo): Option[V]
-    def buildFor(
-                  value: V,
-                  info: ActionsTriggeredInfo,
-                  isIncorrectTeam: Boolean,
-                  botProfile: SlackBotProfile
-                )(implicit request:  Request[AnyContent]): T
-
-    def maybeFor(info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Option[Future[T]] = {
-      maybeValueFor(info).map { v =>
-        buildFor(v, info, botProfile)
-      }
-    }
-
-    def buildFor(value: V, info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Future[T] = {
-      for {
-        maybeUser <- dataService.users.ensureUserFor(LoginInfo(Conversation.SLACK_CONTEXT, info.user.id), botProfile.teamId).map(Some(_))
-        isAdmin <- maybeUser.map { user =>
-          dataService.users.isAdmin(user)
-        }.getOrElse(Future.successful(false))
-        maybeAttemptingSlackTeamIds <- maybeUser.map { user =>
-          dataService.users.maybeSlackTeamIdsFor(user)
-        }.getOrElse(Future.successful(None))
-      } yield {
-        val isSameTeam = maybeAttemptingSlackTeamIds.exists(_.contains(botProfile.slackTeamId))
-        val isIncorrectTeam = !info.team.isEnterpriseGrid && !isSameTeam && !isAdmin
-        buildFor(value, info, isIncorrectTeam, botProfile)
-      }
-    }
-  }
-
-  case class HelpIndexPermission(
-                                  index: Int,
-                                  info: ActionsTriggeredInfo,
-                                  isIncorrectTeam: Boolean,
-                                  botProfile: SlackBotProfile,
-                                  implicit val request: Request[AnyContent]
-                                ) extends HelpPermission {
-
-    val maybeResultText = Some(s"$slackUser clicked More help.")
-
-    def runForCorrectTeam: Unit = {
-      dataService.slackBotProfiles.sendResultWithNewEvent(
-        "help index",
-        (event) => DisplayHelpBehavior(
-          None,
-          None,
-          Some(index),
-          includeNameAndDescription = false,
-          includeNonMatchingResults = false,
-          isFirstTrigger = false,
-          event,
-          services
-        ).result.map(Some(_)),
-        botProfile,
-        info.channel.id,
-        info.user.id,
-        info.message_ts,
-        info.maybeOriginalMessageThreadId,
-        info.isEphemeral,
-        Some(info.response_url),
-        beQuiet = false
-      )
-    }
-
-  }
-
-  object HelpIndexPermission extends HelpPermissionType[HelpIndexPermission, Int] {
-
-    def maybeValueFor(info: ActionsTriggeredInfo): Option[Int] = info.maybeHelpIndexAt
-    def buildFor(
-                  value: Int,
-                  info: ActionsTriggeredInfo,
-                  isIncorrectTeam: Boolean,
-                  botProfile: SlackBotProfile
-                )(implicit request:  Request[AnyContent]): HelpIndexPermission = {
-      HelpIndexPermission(value, info, isIncorrectTeam, botProfile, request)
-    }
-
-  }
-
-  case class HelpForSkillPermission(
-                                     searchValue: HelpGroupSearchValue,
-                                     info: ActionsTriggeredInfo,
-                                     isIncorrectTeam: Boolean,
-                                     botProfile: SlackBotProfile,
-                                     implicit val request: Request[AnyContent]
-                                   ) extends HelpPermission {
-
-    val maybeResultText = Some(info.findButtonLabelForNameAndValue(SHOW_BEHAVIOR_GROUP_HELP, searchValue.helpGroupId).map { text =>
-      s"$slackUser clicked $text."
-    } getOrElse {
-      s"$slackUser clicked a button."
-    })
-
-    def runForCorrectTeam: Unit = {
-      dataService.slackBotProfiles.sendResultWithNewEvent(
-        "skill help with maybe search",
-        (event) => DisplayHelpBehavior(
-          searchValue.maybeSearchText,
-          Some(searchValue.helpGroupId),
-          None,
-          includeNameAndDescription = true,
-          includeNonMatchingResults = false,
-          isFirstTrigger = false,
-          event,
-          services
-        ).result.map(Some(_)),
-        botProfile,
-        info.channel.id,
-        info.user.id,
-        info.message_ts,
-        info.maybeOriginalMessageThreadId,
-        info.isEphemeral,
-        Some(info.response_url),
-        beQuiet = false
-      )
-    }
-
-  }
-
-  object HelpForSkillPermission extends HelpPermissionType[HelpForSkillPermission, HelpGroupSearchValue] {
-
-    def maybeValueFor(info: ActionsTriggeredInfo): Option[HelpGroupSearchValue] = info.maybeHelpForSkillIdWithMaybeSearch
-    def buildFor(
-                  value: HelpGroupSearchValue,
-                  info: ActionsTriggeredInfo,
-                  isIncorrectTeam: Boolean,
-                  botProfile: SlackBotProfile
-                )(implicit request:  Request[AnyContent]): HelpForSkillPermission = {
-      HelpForSkillPermission(value, info, isIncorrectTeam, botProfile, request)
-    }
-
-  }
-
-  case class HelpListAllActionsPermission(
-                                          searchValue: HelpGroupSearchValue,
-                                          info: ActionsTriggeredInfo,
-                                          isIncorrectTeam: Boolean,
-                                          botProfile: SlackBotProfile,
-                                          implicit val request: Request[AnyContent]
-                                          ) extends HelpPermission {
-
-    val maybeResultText = Some(s"$slackUser clicked List all actions")
-
-    def runForCorrectTeam: Unit = {
-      dataService.slackBotProfiles.sendResultWithNewEvent(
-        "for skill action list",
-        event => DisplayHelpBehavior(
-          searchValue.maybeSearchText,
-          Some(searchValue.helpGroupId),
-          None,
-          includeNameAndDescription = false,
-          includeNonMatchingResults = true,
-          isFirstTrigger = false,
-          event,
-          services
-        ).result.map(Some(_)),
-        botProfile,
-        info.channel.id,
-        info.user.id,
-        info.message_ts,
-        info.maybeOriginalMessageThreadId,
-        info.isEphemeral,
-        Some(info.response_url),
-        beQuiet = false
-      )
-    }
-
-  }
-
-  object HelpListAllActionsPermission extends HelpPermissionType[HelpListAllActionsPermission, HelpGroupSearchValue] {
-
-    def maybeValueFor(info: ActionsTriggeredInfo): Option[HelpGroupSearchValue] = info.maybeActionListForSkillId
-    def buildFor(
-                  value: HelpGroupSearchValue,
-                  info: ActionsTriggeredInfo,
-                  isIncorrectTeam: Boolean,
-                  botProfile: SlackBotProfile
-                )(implicit request: Request[AnyContent]): HelpListAllActionsPermission = {
-      HelpListAllActionsPermission(value, info, isIncorrectTeam, botProfile, request)
-    }
-
-  }
-
-  trait ConversationActionPermission extends ActionPermission {
-    val correctUserId: String
-    lazy val isCorrectUser: Boolean = correctUserId == info.user.id
-    lazy val correctUser: String = s"<@$correctUserId>"
-    lazy val shouldRemoveActions: Boolean = isCorrectUser
-
-    def runForCorrectUser(): Unit
-
-    def runInBackground(maybeInstantResponseTs: Future[Option[String]]): Unit = {
-      if (isCorrectUser) {
-        runForCorrectUser()
-      } else {
-        sendEphemeralMessage(s"Only $correctUser can do this", info)
-      }
-    }
-  }
-
-  case class ConfirmContinueConversationPermission(
-                                                    response: ConfirmContinueConversationResponse,
-                                                    info: ActionsTriggeredInfo,
-                                                    implicit val request: Request[AnyContent]
-                                                  ) extends ConversationActionPermission {
-
-    val maybeResultText: Option[String] = {
-      val r = if (response.shouldContinue) { "Yes" } else { "No" }
-      Some(s"$slackUser clicked '$r'")
-    }
-
-    val correctUserId: String = response.userId
-
-    def continue(conversation: Conversation): Future[Unit] = {
-      dataService.conversations.touch(conversation).flatMap { _ =>
-        cacheService.getEvent(conversation.pendingEventKey).map { event =>
-          slackEventService.onEvent(event)
-        }.getOrElse(Future.successful({}))
-      }
-    }
-
-    def dontContinue(conversation: Conversation): Future[Unit] = {
-      dataService.conversations.background(conversation, "OK, on to the next thing.", includeUsername = false).flatMap { _ =>
-        cacheService.getEvent(conversation.pendingEventKey).map { event =>
-          eventHandler.handle(event, None).flatMap { results =>
-            Future.sequence(
-              results.map(result => services.botResultService.sendIn(result, None).map { _ =>
-                Logger.info(event.logTextFor(result, None))
-              })
-            )
-          }.map(_ => {})
-        }.getOrElse(Future.successful({}))
-      }
-    }
-
-    def runForCorrectUser(): Unit = {
-      dataService.conversations.find(response.conversationId).flatMap { maybeConversation =>
-        maybeConversation.map { convo =>
-          if (response.shouldContinue) {
-            continue(convo)
-          } else {
-            dontContinue(convo)
-          }
-        }.getOrElse(Future.successful({}))
-      }
-    }
-  }
-
-  object ConfirmContinueConversationPermission extends ActionPermissionType[ConfirmContinueConversationPermission] {
-
-    def maybeFor(info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Option[Future[ConfirmContinueConversationPermission]] = {
-      info.maybeConfirmContinueConversationResponse.map { response =>
-        buildFor(response, info)
-      }
-    }
-
-    def buildFor(response: ConfirmContinueConversationResponse, info: ActionsTriggeredInfo)(implicit request: Request[AnyContent]): Future[ConfirmContinueConversationPermission] = {
-      Future.successful(ConfirmContinueConversationPermission(
-        response,
-        info,
-        request
-      ))
-    }
-
-  }
-
-  case class StopConversationPermission(
-                                         response: StopConversationResponse,
-                                         info: ActionsTriggeredInfo,
-                                         implicit val request: Request[AnyContent]
-                                       ) extends ConversationActionPermission {
-
-    val correctUserId: String = response.userId
-
-    val maybeResultText = Some(s"$slackUser clicked 'Stop asking'")
-
-    def runForCorrectUser(): Unit = {
-      dataService.conversations.find(response.conversationId).flatMap { maybeConversation =>
-        maybeConversation.map { convo =>
-          dataService.conversations.cancel(convo)
-        }.getOrElse(Future.successful({}))
-      }
-    }
-
-  }
-
-  object StopConversationPermission extends ActionPermissionType[StopConversationPermission] {
-
-    def maybeFor(info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Option[Future[StopConversationPermission]] = {
-      info.maybeStopConversationResponse.map { response =>
-        buildFor(response, info)
-      }
-    }
-
-    def buildFor(response: StopConversationResponse, info: ActionsTriggeredInfo)(implicit request: Request[AnyContent]): Future[StopConversationPermission] = {
-      Future.successful(StopConversationPermission(
-        response,
-        info,
-        request
-      ))
-    }
-
-  }
-
-  case class HelpRunBehaviorVersionPermission(
-                                               behaviorVersionId: String,
-                                               info: ActionsTriggeredInfo,
-                                               isActive: Boolean,
-                                               canBeTriggered: Boolean,
-                                               botProfile: SlackBotProfile,
-                                               implicit val request: Request[AnyContent]
-                                             ) extends ActionPermission {
-
-    val shouldRemoveActions: Boolean = false
-    val maybeOptionLabel: Option[String] = info.findOptionLabelForValue(behaviorVersionId).map(_.mkString("“", "", "”"))
-    val maybeResultText = Some({
-      val actionText = maybeOptionLabel.getOrElse("an action")
-      if (!isActive) {
-        s"$slackUser tried to run an obsolete version of $actionText – run help again to get the latest actions"
-      } else if (!canBeTriggered) {
-        s"$slackUser tried to run $actionText"
-      } else {
-        info.findButtonLabelForNameAndValue(BEHAVIOR_GROUP_HELP_RUN_BEHAVIOR_VERSION, behaviorVersionId).map { text =>
-          s"$slackUser clicked $text"
-        }.getOrElse {
-          s"$slackUser ran $actionText"
-        }
-      }
-    })
-
-    private def runBehaviorVersion(): Unit = {
-      dataService.slackBotProfiles.sendResultWithNewEvent(
-        s"run behavior version $behaviorVersionId",
-        event => for {
-          maybeBehaviorVersion <- dataService.behaviorVersions.findWithoutAccessCheck(behaviorVersionId)
-          maybeResponse <- maybeBehaviorVersion.map { behaviorVersion =>
-            dataService.behaviorResponses.buildFor(
-              event,
-              behaviorVersion,
-              Map(),
-              None,
-              None,
-              None,
-              userExpectsResponse = true
-            ).map(Some(_))
-          }.getOrElse(Future.successful(None))
-          maybeResult <- maybeResponse.map { response =>
-            response.result.map(Some(_))
-          }.getOrElse(Future.successful(None))
-        } yield maybeResult,
-        botProfile,
-        info.channel.id,
-        info.user.id,
-        info.message_ts,
-        info.maybeOriginalMessageThreadId,
-        info.isEphemeral,
-        Some(info.response_url),
-        beQuiet = false
-      )
-    }
-
-    def runInBackground(maybeInstantResponseTs: Future[Option[String]]): Unit = {
-      if (canBeTriggered) {
-        if (isActive) {
-          runBehaviorVersion()
-        }
-      } else {
-        dataService.behaviorVersions.findWithoutAccessCheck(behaviorVersionId).flatMap { maybeBehaviorVersion =>
-          val teamText = maybeBehaviorVersion.map { bv => s" ${bv.team.name}" }.getOrElse("")
-          sendEphemeralMessage(s"Only members of the$teamText team can run this", info)
-        }
-      }
-
-    }
-  }
-
-  object HelpRunBehaviorVersionPermission extends ActionPermissionType[HelpRunBehaviorVersionPermission] {
-
-    def maybeFor(info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Option[Future[HelpRunBehaviorVersionPermission]] = {
-      info.maybeHelpRunBehaviorVersionId.map { behaviorVersionId =>
-        buildFor(behaviorVersionId, info, botProfile)
-      }
-    }
-
-    def buildFor(behaviorVersionId: String, info: ActionsTriggeredInfo, botProfile: SlackBotProfile)(implicit request: Request[AnyContent]): Future[HelpRunBehaviorVersionPermission] = {
-      for {
-        maybeBehaviorVersion <- dataService.behaviorVersions.findWithoutAccessCheck(behaviorVersionId)
-        isActive <- maybeBehaviorVersion.map { behaviorVersion =>
-          dataService.behaviorGroupVersions.isActive(behaviorVersion.groupVersion, Conversation.SLACK_CONTEXT, info.channel.id)
-        }.getOrElse(Future.successful(false))
-        maybeUser <- dataService.users.ensureUserFor(LoginInfo(Conversation.SLACK_CONTEXT, info.user.id), botProfile.teamId).map(Some(_))
-        canBeTriggered <- (for {
-          behaviorVersion <- maybeBehaviorVersion
-          user <- maybeUser
-        } yield behaviorVersion.groupVersion.canBeTriggeredBy(user, dataService)).getOrElse(Future.successful(false))
-      } yield {
-        HelpRunBehaviorVersionPermission(
-          behaviorVersionId,
-          info,
-          isActive = isActive,
-          canBeTriggered = canBeTriggered,
-          botProfile,
-          request
-        )
-      }
-    }
-
   }
 
   def action = Action.async { implicit request =>
@@ -1697,25 +1125,7 @@ class SlackController @Inject() (
             if (info.isValid) {
               info.maybeBotProfile.flatMap { maybeBotProfile =>
                 maybeBotProfile.map { botProfile =>
-                  DataTypeChoicePermission.maybeResultFor(info, botProfile).getOrElse {
-                    YesNoChoicePermission.maybeResultFor(info, botProfile).getOrElse {
-                      ActionChoicePermission.maybeResultFor(info, botProfile).getOrElse {
-                        HelpIndexPermission.maybeResultFor(info, botProfile).getOrElse {
-                          HelpForSkillPermission.maybeResultFor(info, botProfile).getOrElse {
-                            HelpListAllActionsPermission.maybeResultFor(info, botProfile).getOrElse {
-                              ConfirmContinueConversationPermission.maybeResultFor(info, botProfile).getOrElse {
-                                StopConversationPermission.maybeResultFor(info, botProfile).getOrElse {
-                                  HelpRunBehaviorVersionPermission.maybeResultFor(info, botProfile).getOrElse {
-                                    Future.successful(Ok(""))
-                                  }
-                                }
-                              }
-                            }
-                          }
-                        }
-                      }
-                    }
-                  }
+                  permissionResultFor(info, botProfile)
                 }.getOrElse(Future.successful(Ok("")))
               }
             } else {
